refactor: 훈련 세션 WebSocket 폴더 분리
