+++ src/shared/apis/trainingSessions/trainingSessionQueryKeys.ts
@@ -9,9 +9,4 @@ export const trainingSessionQueryKeys = {
   lists: () => [...trainingSessionQueryKeys.all, 'list'] as const,
   // 특정 상태의 세션 목록 쿼리키
   list: (status: TrainingSessionStatus) => [...trainingSessionQueryKeys.lists(), status] as const,
-
-  // 홈 훈련 상태 상세 쿼리의 공통 키
-  statuses: () => [...trainingSessionQueryKeys.all, 'status'] as const,
-  // 세션별 홈 훈련 상태 상세 쿼리키
-  status: (sessionId: string) => [...trainingSessionQueryKeys.statuses(), sessionId] as const,
 };
